rnd: drop range

+++ prog/battleship/ai.c
@@ -105,7 +105,7 @@ int ai_find_orientation(game.state_t *g, state_t *s) {
 	}
 
 	s->d = 0;
-	int n = rnd.intn(navail) + 1;
+	int n = (int) rnd.intn(navail) + 1;
 	while (n) {
 		while (s->used[s->d]) {
 			s->d++;
@@ -240,7 +240,7 @@ void ai_choose_move(game.state_t *g, state_t *s, int *px, int *py) {
 
 	// List preferred possible moves.
 	if (s->ai_turncount++ == 0) {
-		s->ai_huntoffs = rnd.intn(s->ai_srchstep);
+		s->ai_huntoffs = (int) rnd.intn(s->ai_srchstep);
 	}
 	int npref = 0;
 	int ypreferred[BWIDTH * BDEPTH];
@@ -256,11 +256,11 @@ void ai_choose_move(game.state_t *g, state_t *s, int *px, int *py) {
 	}
 
 	if (npref) {
-		int i = rnd.intn(npref);
+		int i = (int) rnd.intn(npref);
 		*px = xpreferred[i];
 		*py = ypreferred[i];
 	} else {
-		int i = rnd.intn(nposs);
+		int i = (int) rnd.intn(nposs);
 		*px = xpossible[i];
 		*py = ypossible[i];
 		if (s->ai_srchstep > 1) {
